renamed files, deleted unused files to improve code coverage
